update required modules

+++ evaluation/slr_eval/python_wer_evaluation.py
@@ -40,12 +40,12 @@ def get_wer_delsubins(ref, hyp, merge_same=False, align_results=False,
     ref_lgt = len(ref) + 1
     hyp_lgt = len(hyp) + 1
 
-    costs = np.ones((ref_lgt, hyp_lgt), dtype=np.int) * 1e6
+    costs = np.ones((ref_lgt, hyp_lgt), dtype=np.int32) * 1e6
     # auxiliary values
     costs[0, :] = np.arange(hyp_lgt) * penalty['ins']
     costs[:, 0] = np.arange(ref_lgt) * penalty['del']
 
-    backtrace = np.zeros((ref_lgt, hyp_lgt), dtype=np.int)
+    backtrace = np.zeros((ref_lgt, hyp_lgt), dtype=np.int32)
     # auxiliary indexes, 0, 1, 2, 3 are corresponding to correct, substitute, insert and delete, respectively
     backtrace[0, :] = 2
     backtrace[:, 0] = 3
